generate acd mix dataset

+++ train.py
@@ -313,7 +313,7 @@ def train_pointmixup(args):
 
         for data in progress_bar:
             rgbs, silhouettes = data['rgb'].to(DEVICE), data['silhouette'].to(DEVICE)
-            points, angles = data['points'].to(DEVICE), data['angle'].to(DEVICE)
+            points, angles = data['points'].to(DEVICE), data['angle'].float().to(DEVICE)
 
             if AUGMENT_3D['rotate']:
                 points = rotate_points_forward_x_axis(points, angles)
